+++ test/basic_crawler.js
@@ -1,5 +1,4 @@
-import chai, { expect } from 'chai';
-import chaiAsPromised from 'chai-as-promised';
+import { expect } from 'chai';
 import _ from 'underscore';
 import sinon from 'sinon';
 import log from 'apify-shared/log';
@@ -9,8 +8,6 @@ import * as Apify from '../build/index';
 import * as keyValueStore from '../build/key_value_store';
 import { RequestQueue, RequestQueueLocal } from '../build/request_queue';
 import { LOCAL_STORAGE_DIR } from './_helper';
-
-chai.use(chaiAsPromised);
 
 describe('BasicCrawler', () => {
     let logLevel;
@@ -580,5 +577,25 @@ describe('BasicCrawler', () => {
         sinon.assert.callCount(addRequestStub, 7);
         expect(count).to.be.eql(7);
         sinon.restore();
+    });
+
+    it('should timeout after handleRequestTimeoutSecs', async () => {
+        const url = 'https://example.com';
+        const requestList = new Apify.RequestList({ sources: [{ url }] });
+        await requestList.initialize();
+
+        const results = [];
+        const crawler = new Apify.BasicCrawler({
+            requestList,
+            handleRequestTimeoutSecs: 0.01,
+            maxRequestRetries: 1,
+            handleRequestFunction: () => delayPromise(1000),
+            handleFailedRequestFunction: ({ request }) => results.push(request),
+        });
+
+        await crawler.run();
+        expect(results).to.have.lengthOf(1);
+        expect(results[0].url).to.be.eql(url);
+        results[0].errorMessages.forEach(msg => expect(msg).to.include('handleRequestFunction timed out'));
     });
 });