+++ test/puppeteer_pool.js
@@ -1,13 +1,10 @@
-import chai, { expect } from 'chai';
+import { expect } from 'chai';
 import fs from 'fs';
-import chaiAsPromised from 'chai-as-promised';
 import _ from 'underscore';
 import log from 'apify-shared/log';
 import { ENV_VARS } from 'apify-shared/consts';
 import * as Apify from '../build/index';
 import { launchPuppeteer } from '../build/puppeteer';
-
-chai.use(chaiAsPromised);
 
 const shortSleep = (millis = 25) => new Promise(resolve => setTimeout(resolve, millis));
 
@@ -535,7 +532,7 @@ describe('PuppeteerPool', () => {
                     await pool.newPage();
                     throw new Error('Invalid error.');
                 } catch (err) {
-                    expect(err.message).to.include('useApifyProxy');
+                    expect(err.stack).to.include('useApifyProxy');
                 }
             });
 
@@ -550,6 +547,32 @@ describe('PuppeteerPool', () => {
                     expect(err.message).to.include('must not be empty');
                 }
             });
+        });
+    });
+
+    describe('prevents hanging of puppeteer operations', () => {
+        let pool;
+        beforeEach(() => {
+            log.setLevel(log.LEVELS.OFF);
+            pool = new Apify.PuppeteerPool({
+                puppeteerOperationTimeoutSecs: 0.05,
+                launchPuppeteerOptions: {
+                    headless: true,
+                },
+            });
+        });
+        afterEach(async () => {
+            log.setLevel(log.LEVELS.ERROR);
+            await pool.destroy();
+        });
+
+        it('should work', async () => {
+            try {
+                await pool._openNewTab(0); // eslint-disable-line no-underscore-dangle
+                throw new Error('invalid error');
+            } catch (err) {
+                expect(err.stack).to.include('timed out');
+            }
         });
     });
 });